+++ SQLiteStudio3/Tests/ParserTest/tst_parsertest.cpp
@@ -26,6 +26,9 @@ class ParserTest : public QObject
         void verifyWindowClause(const QString& sql, SqliteSelectPtr& select, bool& ok);
 
     private Q_SLOTS:
+        void initTestCase();
+        void cleanupTestCase();
+
         void test();
         void testScientificNumber();
         void testUniqConflict();
@@ -56,8 +59,7 @@ class ParserTest : public QObject
         void testWindowClause();
         void testFilterClause();
         void testUpdateFrom();
-        void initTestCase();
-        void cleanupTestCase();
+        void testStringAsTableId();
 };
 
 ParserTest::ParserTest()
@@ -103,7 +105,7 @@ void ParserTest::testGetTableTokens()
     QString sql = "select someTable.* FROM someTable;";
 
     parser3->parse(sql);
-    QVERIFY(parser3->getErrors().size() == 0);
+    QCOMPARE(parser3->getErrors().size(), 0);
 
     SqliteQueryPtr query = parser3->getQueries()[0];
     TokenList tokens = query->getContextTableTokens();
@@ -117,7 +119,7 @@ void ParserTest::testGetTableTokens2()
     QString sql = "select db.tab.col FROM someTable;";
 
     parser3->parse(sql);
-    QVERIFY(parser3->getErrors().size() == 0);
+    QCOMPARE(parser3->getErrors().size(), 0);
 
     SqliteQueryPtr query = parser3->getQueries()[0];
     TokenList tokens = query->getContextTableTokens();
@@ -645,6 +647,14 @@ void ParserTest::testUpdateFrom()
     QVERIFY(update->from->singleSource);
     QVERIFY(update->from->singleSource->select);
     QCOMPARE(update->from->singleSource->alias, "daily");
+}
+
+void ParserTest::testStringAsTableId()
+{
+    QString sql = "select 'bb'.id1 = 'bb'.id2;";
+    bool res = parser3->parse(sql);
+    QVERIFY(res);
+    QVERIFY(parser3->getErrors().isEmpty());
 }
 
 void ParserTest::initTestCase()